made function to store userId
