+++ tapShareFrontend/src/App.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Home from "./pages/Home";
-import SeeFiles from "./pages/SeeFiles";
 import AddCode from "./pages/addCode";
 import ViewCode from "./pages/viewCode";
+import SeeFiles from "./pages/SeeFiles";
 
 const App = () => {
   return (
